#0076 Defining gates and using Gate::denies

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\BlogPost;
 use App\Http\Requests\StorePost;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 // use Illuminate\Support\Facades\DB;
 
@@ -74,12 +75,22 @@ class PostController extends Controller
     public function edit($id)
     {
         $post = BlogPost::findOrFail($id);
+
+        if (Gate::denies('update-post', $post)) {
+            abort(403, "You can't edit this blog post!");
+        }
+
         return view('posts.edit', ['post' => $post]);
     }
 
     public function update(StorePost $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        if (Gate::denies('update-post', $post)) {
+            abort(403, "You can't edit this blog post!");
+        }
+
         $validatedData = $request->validated();
 
         $post->fill($validatedData);
@@ -92,6 +103,11 @@ class PostController extends Controller
     public function destroy(Request $request, $id)
     {
         $post = BlogPost::findOrFail($id);
+
+        if (Gate::denies('delete-post', $post)) {
+            abort(403, "You can't delete this blog post!");
+        }
+
         $post->delete();
 
         // BlogPost::destroy($id);
